regroup slice packages, add a few more funcs
