+++ slices/slice_test.go
@@ -8,32 +8,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestAny(t *testing.T) {
-	f := func(given []int, expected bool) {
-		even := func(t int) bool { return (t % 2) == 0 }
-		actual := slices.Any(given, even)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, false)
-	f([]int{1, 3}, false)
-	f([]int{2}, true)
-	f([]int{1, 2}, true)
-}
-
-func TestAll(t *testing.T) {
-	f := func(given []int, expected bool) {
-		even := func(t int) bool { return (t % 2) == 0 }
-		actual := slices.All(given, even)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, true)
-	f([]int{2}, true)
-	f([]int{1}, false)
-	f([]int{2, 4}, true)
-	f([]int{2, 4, 1}, false)
-	f([]int{1, 2, 4}, false)
-}
-
 func TestChoice(t *testing.T) {
 	f := func(given []int, seed int64, expected int) {
 		actual, _ := slices.Choice(given, seed)
@@ -42,18 +16,6 @@ func TestChoice(t *testing.T) {
 	f([]int{1}, 0, 1)
 	f([]int{1, 2, 3}, 1, 3)
 	f([]int{1, 2, 3}, 2, 2)
-}
-
-func TestChunkBy(t *testing.T) {
-	f := func(given []int, expected [][]int) {
-		reminder := func(t int) int { return (t % 2) }
-		actual := slices.ChunkBy(given, reminder)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, [][]int{})
-	f([]int{1}, [][]int{{1}})
-	f([]int{1, 2, 3}, [][]int{{1}, {2}, {3}})
-	f([]int{1, 3, 2, 4, 5}, [][]int{{1, 3}, {2, 4}, {5}})
 }
 
 func TestChunkEvery(t *testing.T) {
@@ -94,19 +56,6 @@ func TestCount(t *testing.T) {
 	f(1, []int{1, 1, 1, 1, 1}, 5)
 }
 
-func TestCountFunc(t *testing.T) {
-	f := func(given []int, expected int) {
-		even := func(t int) bool { return (t % 2) == 0 }
-		actual := slices.CountFunc(given, even)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, 0)
-	f([]int{1}, 0)
-	f([]int{2}, 1)
-	f([]int{1, 2, 3, 4, 5}, 2)
-	f([]int{1, 2, 3, 4, 5, 6}, 3)
-}
-
 func TestCycle(t *testing.T) {
 	f := func(count int, given []int, expected []int) {
 		c := slices.Cycle(given)
@@ -131,21 +80,6 @@ func TestDedup(t *testing.T) {
 	f([]int{1, 2, 3}, []int{1, 2, 3})
 	f([]int{1, 2, 2, 3}, []int{1, 2, 3})
 	f([]int{1, 2, 2, 3, 3, 3, 2, 1, 1}, []int{1, 2, 3, 2, 1})
-}
-
-func TestDedupBy(t *testing.T) {
-	f := func(given []int, expected []int) {
-		even := func(el int) int { return el % 2 }
-		actual := slices.DedupBy(given, even)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, []int{})
-	f([]int{1}, []int{1})
-	f([]int{1, 1}, []int{1})
-	f([]int{1, 2}, []int{1, 2})
-	f([]int{1, 2, 3}, []int{1, 2, 3})
-	f([]int{1, 2, 2, 3}, []int{1, 2, 3})
-	f([]int{1, 2, 4, 3, 5, 7, 10}, []int{1, 2, 3, 10})
 }
 
 func TestDelete(t *testing.T) {
@@ -210,21 +144,6 @@ func TestDropEvery(t *testing.T) {
 	f([]int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}, 3, 1, []int{1, 2, 4, 5, 7, 8, 10})
 }
 
-func TestDropWhile(t *testing.T) {
-	f := func(given []int, expected []int) {
-		even := func(el int) bool { return el%2 == 0 }
-		actual := slices.DropWhile(given, even)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, []int{})
-	f([]int{2}, []int{})
-	f([]int{1}, []int{1})
-	f([]int{2, 1}, []int{1})
-	f([]int{2, 1, 2}, []int{1, 2})
-	f([]int{1, 2}, []int{1, 2})
-	f([]int{2, 4, 6, 1, 8}, []int{1, 8})
-}
-
 func TestEndsWith(t *testing.T) {
 	f := func(given []int, suffix []int, expected bool) {
 		actual := slices.EndsWith(given, suffix)
@@ -267,50 +186,6 @@ func TestEqual(t *testing.T) {
 	f([]int{1, 2, 3}, []int{}, false)
 }
 
-func TestFilter(t *testing.T) {
-	f := func(given []int, expected []int) {
-		even := func(t int) bool { return (t % 2) == 0 }
-		actual := slices.Filter(given, even)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, []int{})
-	f([]int{1, 2, 3, 4}, []int{2, 4})
-	f([]int{1, 3}, []int{})
-	f([]int{2, 4}, []int{2, 4})
-}
-
-func TestFind(t *testing.T) {
-	f := func(given []int, expectedEl int, expectedErr error) {
-		even := func(t int) bool { return (t % 2) == 0 }
-		el, err := slices.Find(given, even)
-		require.Equal(t, expectedEl, el)
-		require.Equal(t, expectedErr, err)
-	}
-	f([]int{}, 0, slices.ErrNotFound)
-	f([]int{1}, 0, slices.ErrNotFound)
-	f([]int{1}, 0, slices.ErrNotFound)
-	f([]int{2}, 2, nil)
-	f([]int{1, 2}, 2, nil)
-	f([]int{1, 2, 3}, 2, nil)
-	f([]int{1, 3, 5}, 0, slices.ErrNotFound)
-}
-
-func TestFindIndex(t *testing.T) {
-	f := func(given []int, expectedInd int) {
-		even := func(t int) bool { return (t % 2) == 0 }
-		index := slices.FindIndex(given, even)
-		require.Equal(t, expectedInd, index)
-	}
-	f([]int{}, -1)
-	f([]int{1}, -1)
-	f([]int{1}, -1)
-	f([]int{2}, 0)
-	f([]int{1, 2}, 1)
-	f([]int{1, 2, 3}, 1)
-	f([]int{1, 3, 5, 7, 9, 2}, 5)
-	f([]int{1, 3, 5}, -1)
-}
-
 func TestJoin(t *testing.T) {
 	is := require.New(t)
 	is.Equal(slices.Join([]int{}, ""), "")
@@ -323,17 +198,6 @@ func TestJoin(t *testing.T) {
 	is.Equal(slices.Join([]int{1, 2, 3}, "|"), "1|2|3")
 	is.Equal(slices.Join([]int{1, 2, 3}, "<T>"), "1<T>2<T>3")
 	is.Equal(slices.Join([]string{"hello", "world"}, " "), "hello world")
-}
-
-func TestGroupBy(t *testing.T) {
-	f := func(given []int, expected map[int][]int) {
-		reminder := func(t int) int { return (t % 2) }
-		actual := slices.GroupBy(given, reminder)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, map[int][]int{})
-	f([]int{1}, map[int][]int{1: {1}})
-	f([]int{1, 3, 2, 4, 5}, map[int][]int{0: {2, 4}, 1: {1, 3, 5}})
 }
 
 func TestInsertAt(t *testing.T) {
@@ -373,17 +237,6 @@ func TestLast(t *testing.T) {
 	f([]int{}, 0, slices.ErrEmpty)
 	f([]int{1}, 1, nil)
 	f([]int{1, 2, 3}, 3, nil)
-}
-
-func TestMap(t *testing.T) {
-	f := func(given []int, expected []int) {
-		double := func(t int) int { return (t * 2) }
-		actual := slices.Map(given, double)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, []int{})
-	f([]int{1}, []int{2})
-	f([]int{1, 2, 3}, []int{2, 4, 6})
 }
 
 func TestMax(t *testing.T) {
@@ -456,36 +309,6 @@ func TestProduct(t *testing.T) {
 	})
 }
 
-func TestReduce(t *testing.T) {
-	f := func(given []int, expected int) {
-		sum := func(el int, acc int) int { return (el) + acc }
-		actual := slices.Reduce(given, 0, sum)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, 0)
-	f([]int{1}, 1)
-	f([]int{1, 2}, 3)
-	f([]int{1, 2, 3}, 6)
-}
-
-func TestReduceWhile(t *testing.T) {
-	f := func(given []int, expected int) {
-		sum := func(el int, acc int) (int, error) {
-			if el == 0 {
-				return acc, slices.ErrEmpty
-			}
-			return (el) + acc, nil
-		}
-		actual, _ := slices.ReduceWhile(given, 0, sum)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, 0)
-	f([]int{1}, 1)
-	f([]int{1, 2}, 3)
-	f([]int{1, 2, 3}, 6)
-	f([]int{1, 2, 0, 3}, 3)
-}
-
 func TestReverse(t *testing.T) {
 	f := func(given []int, expected []int) {
 		actual := slices.Reverse(given)
@@ -526,19 +349,6 @@ func TestSame(t *testing.T) {
 	f([]int{1, 2, 1}, false)
 	f([]int{1, 2, 2}, false)
 	f([]int{1, 1, 2}, false)
-}
-
-func TestScan(t *testing.T) {
-	f := func(given []int, expected []int) {
-		sum := func(el int, acc int) int { return (el) + acc }
-		actual := slices.Scan(given, 0, sum)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, []int{})
-	f([]int{1}, []int{1})
-	f([]int{1, 2}, []int{1, 3})
-	f([]int{1, 2, 3}, []int{1, 3, 6})
-	f([]int{1, 2, 3, 4}, []int{1, 3, 6, 10})
 }
 
 func TestShuffle(t *testing.T) {
@@ -639,19 +449,6 @@ func TestTakeRandom(t *testing.T) {
 	f([]int{1}, 1, 0, []int{1})
 	f([]int{1, 2, 3, 4, 5}, 3, 1, []int{3, 1, 2})
 	f([]int{1, 2, 3, 4, 5}, 5, 1, []int{3, 1, 2, 5, 4})
-}
-
-func TestTakeWhile(t *testing.T) {
-	f := func(given []int, expected []int) {
-		even := func(el int) bool { return el%2 == 0 }
-		actual := slices.TakeWhile(given, even)
-		require.Equal(t, expected, actual)
-	}
-	f([]int{}, []int{})
-	f([]int{1}, []int{})
-	f([]int{2}, []int{2})
-	f([]int{2, 4, 6, 1, 8}, []int{2, 4, 6})
-	f([]int{1, 2, 3}, []int{})
 }
 
 func TestUniq(t *testing.T) {
